rplugin/python3/qualia/requirements.txt
git+https://github.com/neovim/pynvim
# pynvim


markdown-it-py[plugins]
lmdb
appdirs
orderedset
cryptography
base65536
bloomfilter-py

typing-extensions

firebase-admin
urllib3
pid
ntplib
requests
psutil

rpyc

# Dev
typeguard

